+++ .dockerignore
@@ -1,7 +1,8 @@
-/resources
 /assets
 /logs
+/tests
+/resources
 
 # Builds
 *.test
-go-cdn
+./go-cdn
