+++ configs/jaeger.json.sample
