+++ internal/database/caching.go
@@ -3,10 +3,10 @@ package database
 import (
 	"context"
 	"fmt"
-	"go-cdn/config"
-	"go-cdn/consul"
-	"go-cdn/tracing"
-	"go-cdn/utils"
+	"go-cdn/internal/config"
+	"go-cdn/internal/consul"
+	"go-cdn/internal/tracing"
+	"go-cdn/pkg/utils"
 	"log"
 	"strconv"
 	"strings"
