+++ internal/server/server.go
@@ -3,10 +3,10 @@ package server
 import (
 	"context"
 	"fmt"
-	"go-cdn/config"
-	"go-cdn/database"
-	"go-cdn/tracing"
-	"go-cdn/utils"
+	"go-cdn/internal/config"
+	"go-cdn/internal/database"
+	"go-cdn/internal/tracing"
+	"go-cdn/pkg/utils"
 	"io"
 	"net/http"
 	"os/signal"
