+++ k8s/cdn-deployment.sample.yml
