+++ k8s/cdn-service.sample.yml
