+++ k8s/tracing/jaeger-namespace.yml
